WIP: Inspector refactor
