+++ lib/engine/include/facade/engine/editor/inspector.hpp
@@ -1,82 +1,20 @@
 #pragma once
 #include <facade/engine/editor/common.hpp>
-#include <facade/scene/material.hpp>
-#include <facade/scene/node.hpp>
-#include <facade/scene/scene_resources.hpp>
-#include <facade/util/nvec3.hpp>
-#include <facade/util/rgb.hpp>
-#include <limits>
+#include <facade/scene/scene.hpp>
 
-namespace facade {
-struct Lights;
-class Scene;
-
-namespace editor {
-///
-/// \brief Stateless ImGui helper to inspect / reflect various properties
-///
+namespace facade::editor {
 class Inspector {
   public:
-	static constexpr auto min_v{-std::numeric_limits<float>::max()};
-	static constexpr auto max_v{std::numeric_limits<float>::max()};
+	Inspector(NotClosed<Window> target, Scene& out_scene) : m_target(target), m_scene(out_scene) {}
 
-	///
-	/// \brief Construct an Inspector instance
-	///
-	/// Inspectors don't do anything on construction, constructors exist to enforce invariants instance-wide.
-	/// For all Inspectors, an existing Window target is required, Inspector instances will not create any
-	///
-	Inspector(NotClosed<Window>) {}
-
-	bool inspect(char const* label, glm::vec2& out_vec2, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
-	bool inspect(char const* label, glm::vec3& out_vec3, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
-	bool inspect(char const* label, glm::vec4& out_vec4, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
-	bool inspect(char const* label, nvec3& out_vec3, float speed = 0.01f) const;
-	bool inspect(char const* label, glm::quat& out_quat) const;
-	bool inspect_rgb(char const* label, glm::vec3& out_rgb) const;
-	bool inspect(char const* label, Rgb& out_rgb) const;
-	bool inspect(Transform& out_transform, Bool& out_unified_scaling) const;
-	bool inspect(std::span<Transform> out_instances, Bool unfied_scaling) const;
-	bool inspect(Lights& out_lights) const;
+	void resources() const;
+	void transform(Node& out_node, Bool& out_unified_scaling) const;
+	void instances(Node& out_node, Bool unified_scaling) const;
+	void mesh(Node& out_node) const;
+	void node(Id<Node> node_id, Bool& out_unified_scaling) const;
 
   private:
-	bool do_inspect(Transform& out_transform, Bool& out_unified_scaling, Bool scaling_toggle) const;
-};
-
-class SceneInspector : public Inspector {
-  public:
-	using Inspector::inspect;
-
-	SceneInspector(NotClosed<Window> target, Scene& scene);
-
-	bool inspect(NotClosed<TreeNode>, UnlitMaterial& out_material) const;
-	bool inspect(NotClosed<TreeNode> node, LitMaterial& out_material) const;
-	bool inspect(Id<Material> material_id) const;
-	bool inspect(Id<Mesh> mesh_id) const;
-	bool inspect(Id<Camera> camera_id) const;
-
-	bool inspect(Id<Node> node_id, Bool& out_unified_scaling) const;
-
-  private:
-	Scene& m_scene;
 	NotClosed<Window> m_target;
+	Scene& m_scene;
 };
-
-class ResourceInspector {
-  public:
-	ResourceInspector(NotClosed<Window>, SceneResources const& resources);
-
-	void display() const;
-	void display(Camera const& camera, std::size_t index, std::string_view prefix = {}) const;
-	void display(Texture const& texture, std::size_t index, std::string_view prefix = {}) const;
-	void display(Material const& material, std::size_t const index, std::string_view prefix = {}) const;
-	void display(Mesh const& mesh, std::size_t const index, std::string_view prefix = {}) const;
-
-  private:
-	void display(LitMaterial const& lit) const;
-	void display(UnlitMaterial const& unlit) const;
-
-	SceneResources const& m_resources;
-};
-} // namespace editor
-} // namespace facade
+} // namespace facade::editor
