+++ lib/engine/include/facade/engine/editor/reflector.hpp
@@ -1,0 +1,40 @@
+#pragma once
+#include <facade/engine/editor/common.hpp>
+#include <facade/scene/camera.hpp>
+#include <facade/util/nvec3.hpp>
+#include <facade/util/rgb.hpp>
+#include <facade/util/transform.hpp>
+#include <glm/gtc/quaternion.hpp>
+
+namespace facade::editor {
+///
+/// \brief Stateless ImGui helper to reflect various properties
+///
+class Reflector {
+  public:
+	struct AsRgb {
+		glm::vec3& out;
+	};
+
+	static constexpr auto min_v{-std::numeric_limits<float>::max()};
+	static constexpr auto max_v{std::numeric_limits<float>::max()};
+
+	///
+	/// \brief Construct an Reflector instance
+	///
+	/// Reflectors don't do anything on construction, constructors exist to enforce invariants instance-wide.
+	/// For all Reflectors, an existing Window target is required, Reflector instances will not create any
+	///
+	Reflector(NotClosed<Window>) {}
+
+	bool operator()(char const* label, glm::vec2& out_vec2, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
+	bool operator()(char const* label, glm::vec3& out_vec3, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
+	bool operator()(char const* label, glm::vec4& out_vec4, float speed = 1.0f, float lo = min_v, float hi = max_v) const;
+	bool operator()(char const* label, nvec3& out_vec3, float speed = 0.01f) const;
+	bool operator()(char const* label, glm::quat& out_quat) const;
+	bool operator()(char const* label, AsRgb out_rgb) const;
+	bool operator()(Rgb& out_rgb) const;
+	bool operator()(Transform& out_transform, Bool& out_unified_scaling, Bool scaling_toggle) const;
+	bool operator()(Camera& out_camera) const;
+};
+} // namespace facade::editor
