+++ lib/engine/src/editor/common.cpp
@@ -13,6 +13,16 @@ glm::vec2 max_size(std::span<char const* const> strings) {
 	return ret;
 }
 
+bool small_button_red(char const* label) {
+	bool ret = false;
+	ImGui::PushStyleColor(ImGuiCol_Button, {0.8f, 0.0f, 0.1f, 1.0f});
+	ImGui::PushStyleColor(ImGuiCol_ButtonHovered, {1.0f, 0.0f, 0.1f, 1.0f});
+	ImGui::PushStyleColor(ImGuiCol_ButtonActive, {0.6f, 0.0f, 0.1f, 1.0f});
+	if (ImGui::SmallButton(label)) { ret = true; }
+	ImGui::PopStyleColor(3);
+	return ret;
+}
+
 Openable::Openable(bool is_open) : m_open(is_open) {}
 
 Window::Window(char const* label, bool* open_if, int flags) : Canvas(ImGui::Begin(label, open_if, flags)) {}
