+++ src/gui/main_menu.cpp
@@ -49,19 +49,20 @@ bool WindowMenu::display_scene(Scene& scene) {
 	return show;
 }
 
-bool WindowMenu::display_lights(Lights& lights) {
+bool WindowMenu::display_lights(Lights&) {
 	ImGui::SetNextWindowSize({400.0f, 400.0f}, ImGuiCond_FirstUseEver);
 	bool show{true};
-	if (auto window = editor::Window{"Lights", &show}) { editor::Inspector{window}.inspect(lights); }
+	if (auto window = editor::Window{"Lights", &show}) {
+		// editor::Inspector{window}.inspect(lights);
+		// TODO
+	}
 	return show;
 }
 
 bool WindowMenu::display_inspector(Scene& scene) {
 	bool show = true;
 	ImGui::SetNextWindowSize({400.0f, 400.0f}, ImGuiCond_FirstUseEver);
-	if (auto window = editor::Window{m_data.inspect.name.c_str(), &show}) {
-		editor::SceneInspector{window, scene}.inspect(m_data.inspect.id, m_data.unified_scaling);
-	}
+	if (auto window = editor::Window{m_data.inspect.name.c_str(), &show}) { editor::Inspector{window, scene}.node(m_data.inspect.id, m_data.unified_scaling); }
 	return m_data.inspect && show;
 }
 
@@ -105,7 +106,7 @@ bool WindowMenu::display_log() {
 bool WindowMenu::display_resources(Scene& scene) {
 	bool show{true};
 	ImGui::SetNextWindowSize({600.0f, 200.0f}, ImGuiCond_FirstUseEver);
-	if (auto window = editor::Window{"Resources", &show}) { editor::ResourceInspector{window, scene.resources()}.display(); }
+	if (auto window = editor::Window{"Resources", &show}) { editor::Inspector{window, scene}.resources(); }
 	return show;
 }
